refactor : FileUtil 을 FileRepository 로 변경
- File 을 저장하는 것으로 보고 Repository 라고 판단하여 infra 계층으로 변경
- Local, S3 로 FileRepository 를 구현하여 사용

관련 이슈
- #117
